Organize and standardize testing mocks for executor and privilege packages
- Rename executor testing files into testing subpackage (mocks.go and
  testify_mocks.go) and update package names to testing for clearer
  separation between production and test helpers.
- Update executor tests to import the new testing package and replace
  direct executor.Mock* usages with executortesting equivalents.
- Rename privilege testing mock file into testing/mocks.go and add
  appropriate build tag to ensure correct test-only inclusion.
- Add Mock File Organization section to CLAUDE.md documenting the
  conventions for mock placement, file naming, and usage.
